+++ src/Controls/src/Core/HandlerImpl/View.Android.cs
@@ -1,9 +1,6 @@
 ﻿#nullable enable
 
-using AndroidX.Core.View;
-using AndroidX.Core.View.Accessibility;
 using Microsoft.Maui.Controls.Platform;
-using NativeView = Android.Views.View;
 
 namespace Microsoft.Maui.Controls
 {
@@ -12,24 +9,6 @@ namespace Microsoft.Maui.Controls
 		partial void HandlerChangedPartial()
 		{
 			this.ApplyControlsAccessibilityDelegateIfNeeded();
-		}
-	}
-
-	class ControlsAccessibilityDelegate : AccessibilityDelegateCompatWrapper
-	{
-		public IViewHandler Handler { get; }
-
-		public ControlsAccessibilityDelegate(AccessibilityDelegateCompat? originalDelegate, IViewHandler viewHandler) : base(originalDelegate)
-		{
-			Handler = viewHandler;
-		}
-
-		public override void OnInitializeAccessibilityNodeInfo(NativeView? host, AccessibilityNodeInfoCompat? info)
-		{
-			base.OnInitializeAccessibilityNodeInfo(host, info);
-
-			if (Handler?.NativeView is NativeView nativeView && Handler?.VirtualView is View v)
-				v.UpdateSemanticNodeInfo(info);
 		}
 	}
 }
